pyproject.toml
[tool.poetry]
name = "data-engine-poc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "data_engine_poc"}]

[tool.poetry.dependencies]
python = ">=3.10.14"
psycopg2-binary = "^2.9.9"
faker = "^30.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
